+++ middleware/logging/log.go
@@ -9,10 +9,13 @@ package logging
  */
 
 import (
+	"bytes"
 	"context"
+	"io"
 	"net/http"
 	"time"
 
+	"github.com/Southclaws/fault"
 	"github.com/google/uuid"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
@@ -30,11 +33,15 @@ func (mw MW) Wrap(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		start := time.Now()
 		reqID := uuid.New().String()
+		var path string
+		if r.URL != nil {
+			path = r.URL.Path
+		}
 		contextLogger := log.With().
 			Str("reqID", reqID).
 			Str("remoteAddr", r.RemoteAddr).
 			Str("method", r.Method).
-			Str("path", r.URL.Path).
+			Str("path", path).
 			Str("host", r.Host).
 			Bool("tls", r.TLS != nil).
 			Logger()
@@ -45,12 +52,14 @@ func (mw MW) Wrap(next http.Handler) http.Handler {
 		r = r.WithContext(ctx)
 
 		if contextLogger.GetLevel() <= zerolog.TraceLevel {
+			request, err := PeekBody(r)
+			if err != nil {
+				w.WriteHeader(http.StatusInternalServerError)
+				return
+			}
 
 			myResponseWriter := newMyResponseWriter(w)
-			myRequestReader := newMyRequestReader(r)
 			next.ServeHTTP(myResponseWriter, r)
-
-			request := myRequestReader.requestBody.Bytes()
 
 			contextLogger.Trace().
 				Bytes("request", request).
@@ -76,4 +85,20 @@ func LoggerFromCtx(ctx context.Context) zerolog.Logger {
 		return log
 	}
 	return log.Logger
+}
+
+func PeekBody(r *http.Request) ([]byte, error) {
+	// consume request body
+	var request []byte
+	if r.Body != nil {
+		var err error
+		request, err = io.ReadAll(r.Body)
+		if err != nil {
+			return nil, fault.Wrap(err)
+		}
+		_ = r.Body.Close()
+		// restore original body for other middlewares
+		r.Body = io.NopCloser(bytes.NewBuffer(request))
+	}
+	return request, nil
 }